fix: - Circle dependency - Updated peerDependencies
