+++ package.json
@@ -27,7 +27,9 @@
     "chromatic": "npx chromatic --project-token dqq8k95raoo"
   },
   "peerDependencies": {
-    "react": "^16.0.0"
+    "rebass": "^4.0.7",
+    "react": "^16.13.1",
+    "emotion-theming": "^10.0.27"
   },
   "devDependencies": {
     "@babel/core": "^7.11.4",
@@ -67,7 +69,6 @@
     "microbundle-crl": "^0.13.10",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.4",
-    "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-is": "^16.13.1",
     "react-scripts": "^3.4.1",
@@ -82,8 +83,6 @@
     "@fortawesome/free-regular-svg-icons": "^5.14.0",
     "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.11",
-    "@styled-system/css": "^5.1.5",
-    "emotion-theming": "^10.0.27",
-    "rebass": "^4.0.7"
+    "@styled-system/css": "^5.1.5"
   }
 }
