+++ src/components/navigation/types.ts
@@ -1,5 +1,17 @@
 import { IconProp } from '@fortawesome/fontawesome-svg-core';
-import { NavigationItemProps } from './item';
+import React from 'react';
+import { BoxProps } from 'rebass';
+
+export interface NavigationItemProps extends Omit<BoxProps, 'css'> {
+  title: string;
+  icon?: IconProp;
+  isActive?: boolean;
+  isSubItem?: boolean;
+  hasDivider?: boolean;
+  children?: React.ReactNode;
+  disabled?: boolean;
+  onClick?: () => void;
+}
 
 export interface TreeNode extends NavigationItemProps {
   title: string;
